refactor(cli): use command builder
