+++ renku/cli/rerun.py
@@ -47,11 +47,10 @@ from pathlib import Path
 
 import click
 
-from renku.cli.update import execute_workflow
+from renku.cli.utils.callback import ClickCallback
 from renku.core import errors
-from renku.core.commands.client import pass_local_client
-from renku.core.commands.graph import Graph
 from renku.core.commands.options import option_siblings
+from renku.core.commands.rerun import rerun_workflows
 
 
 def show_inputs(client, workflow):
@@ -95,36 +94,13 @@ def edit_inputs(client, workflow):
 @click.option(
     "--default-inputs", "inputs", default=True, flag_value=lambda _, workflow: workflow, help="Use default inputs.",
 )
-@click.option(
-    "--show-inputs", "inputs", flag_value=show_inputs, help=show_inputs.__doc__,
-)
-@click.option(
-    "--edit-inputs", "inputs", flag_value=edit_inputs, help=edit_inputs.__doc__,
-)
-@click.argument(
-    "paths", type=click.Path(exists=True, dir_okay=True), nargs=-1, required=True,
-)
-@pass_local_client(
-    clean=True, requires_migration=True, commit=True,
-)
-def rerun(client, revision, roots, siblings, inputs, paths):
+@click.option("--show-inputs", "inputs", flag_value=show_inputs, help=show_inputs.__doc__)
+@click.option("--edit-inputs", "inputs", flag_value=edit_inputs, help=edit_inputs.__doc__)
+@click.argument("paths", type=click.Path(exists=True, dir_okay=True), nargs=-1, required=True)
+def rerun(revision, roots, siblings, inputs, paths):
     """Recreate files generated by a sequence of ``run`` commands."""
-    graph = Graph(client)
-    outputs = graph.build(paths=paths, revision=revision)
+    communicator = ClickCallback()
 
-    # Check or extend siblings of outputs.
-    outputs = siblings(graph, outputs)
-    output_paths = {node.path for node in outputs}
-
-    # Normalize and check all starting paths.
-    roots = {graph.normalize_path(root) for root in roots}
-    output_paths -= roots
-    outputs = [o for o in outputs if o.path not in roots]
-
-    # Generate workflow and check inputs.
-    # NOTE The workflow creation is done before opening a new file.
-    workflow = inputs(client, graph.as_workflow(input_paths=roots, output_paths=output_paths, outputs=outputs,))
-
-    execute_workflow(
-        client=client, workflow=workflow, output_paths=output_paths, command_name="rerun", update_commits=False
+    rerun_workflows().with_communicator(communicator).build().execute(
+        revision=revision, roots=roots, siblings=siblings, inputs=inputs, paths=paths
     )
